remove unnecessary useEffect, merge jumpToHash into its only used location
